lesson 5 completed

+++ src/services/user.service.ts
@@ -1,21 +1,63 @@
-import { IUser, IUserDTO } from "../interfaces/user.interface";
+import { StatusCodesEnum } from "../enums/status-codes.enum";
+import { ApiError } from "../errors/api.error";
+import {
+    IUser,
+    IUserCreateDTO,
+    IUserUpdateDTO,
+} from "../interfaces/user.interface";
 import { userRepository } from "../repositories/user.repository";
 
 class UserService {
     public getAll(): Promise<IUser[]> {
         return userRepository.getAll();
     }
-    public create(user: IUserDTO): Promise<IUser> {
+
+    public create(user: IUserCreateDTO): Promise<IUser> {
         return userRepository.create(user);
     }
-    public getById(userId: string): Promise<IUserDTO> {
-        return userRepository.getById(userId);
+
+    public async getById(userId: string): Promise<IUser> {
+        const user = await userRepository.getById(userId);
+
+        if (!user) {
+            throw new ApiError("User not found", StatusCodesEnum.NOT_FOUND);
+        }
+
+        return user;
     }
-    public updateById(userId: string, newUser: IUserDTO): Promise<IUser> {
-        return userRepository.updateById(userId, newUser);
+
+    public async updateById(
+        userId: string,
+        updateData: IUserUpdateDTO,
+    ): Promise<IUser> {
+        const user = await userRepository.getById(userId);
+
+        if (!user) {
+            throw new ApiError("User not found", StatusCodesEnum.NOT_FOUND);
+        }
+
+        return await userRepository.updateById(userId, updateData);
     }
-    public deleteById(userId: string) {
-        return userRepository.deleteById(userId);
+
+    public async deleteById(userId: string): Promise<void> {
+        const user = await userRepository.getById(userId);
+
+        if (!user) {
+            throw new ApiError("User not found", StatusCodesEnum.NOT_FOUND);
+        }
+
+        await userRepository.deleteById(userId);
+    }
+
+    public async isEmailUnique(email: string): Promise<void> {
+        const user = await userRepository.getByEmail(email);
+
+        if (user) {
+            throw new ApiError(
+                "User is already exists",
+                StatusCodesEnum.BAD_REQUEST,
+            );
+        }
     }
 }
